+++ node/resolvers/index.ts
@@ -4,8 +4,8 @@ import { fieldResolvers as catalogFieldResolvers, queries as catalogQueries } fr
 import { fieldResolvers as checkoutFieldResolvers, mutations as checkoutMutations, queries as checkoutQueries } from './checkout'
 import { mutations as documentMutations, queries as documentQueries } from './document'
 import { mutation as listMutations, queries as listQueries } from './list'
-import { queries as logisticsQueries } from './logistics'
-import { fieldResolvers as profileFieldResolvers, mutations as profileMutations, queries as profileQueries} from './profile'
+import { fieldResolvers as logisticsResolvers, queries as logisticsQueries } from './logistics'
+import { mutations as profileMutations, queries as profileQueries, fieldResolvers as profileFieldResolvers } from './profile'
 import { mutations as sessionMutations, queries as sessionQueries } from './session'
 import { fieldResolvers as subscriptionsFieldResolvers, mutations as subscriptionsMutations, queries as subscriptionsQueries } from './subscriptions'
 
@@ -18,6 +18,7 @@ export const resolvers = {
   ...profileFieldResolvers,
   ...checkoutFieldResolvers,
   ...subscriptionsFieldResolvers,
+  ...logisticsResolvers,
   Mutation: {
     ...profileMutations,
     ...checkoutMutations,